requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.7
# by the following command:
#
#    pip-compile requirements.in
#
attrs==23.1.0
    # via -r requirements.in
h5py==3.8.0
    # via -r requirements.in
importlib-metadata==6.4.1
    # via attrs
numpy==1.21.6
    # via
    #   -r requirements.in
    #   h5py
typing-extensions==4.5.0
    # via importlib-metadata
zipp==3.15.0
    # via importlib-metadata
